Fix: 'nomalmap' typo (cats-blender-plugin will support correct spell)
In the past, this can't be fixed because 'cats-blender-plugin' used
'nomalmapTexture'(https://github.com/GiveMeAllYourCats/cats-blender-plugin/commit/a488049d5ed90480965a40116d6bb37fcfda4515)
But on the develop branch, https://github.com/GiveMeAllYourCats/cats-blender-plugin/commit/c6aead4783abff1adaa0744aebd1b15d990066f8 fixed
it to support correct spell.
Because cats-blender-plugin now support both 'nomalmap' and
'normalmap' in the **Next release**, we can prepare to fix this typo.

Note that it **isn't pushed to master** yet at this moment, so that this
can't be in release for now