Fix for exporting VRM model imported with older addon
Support models that were loaded by earlier versions(1_13_1 or earlier,
I don't know correctly, which had this typo

Those models have node.inputs["NomalmapTexture"] instead of
"NormalmapTexture".
But 'shader_val', which is come from
MaterialMtoon.texture_kind_exchange_dic, can be "NormalmapTexture".
if script reference node.inputs["NormalmapTexture"] in that situation,
it will occur error.
So change it to "NomalmapTexture" which is typo but points to the same
thing in those models.